refactor: Move cost center policy to financial repo
